Bringing in branch develop.
This commit brings in the work on relaxing to prescribed barotrope.
There are a few details where this "outer loop" convergence differs
form the equivalent one in CMS. The pressures and densities are
interpreted as being on the same grid point, rather than surfaces
or interiors of layers. Ultimately this is not that important though.
It turned out to be more precise to use density midpoints (by linear
interpolation) when calculating pressure by integration of the
hydrostatic equation. It also turned out that using a continuously
changing total mass when converting from dimensionless potential (A0)
to physical potential (in order to use an eos) was confusing the last
little bit of density convergence. Using a fixed, reference mass (like
we do in CMSPlanet) solves that issue.

As always, we converge not to the input barotrope but to a scaled
barotrope, if we want to match total mass. Here we also want to
match a reference equatorial radius. If beta is the mass factor and
alpha is the radius factor (ref-over-calc) then the converged barotrope
actually follows P=beta/alpha*f(rho/beta).