Level distribution strategy and tests.
Testing of level distributions suggests we can't really do much better
than equal-spaced radii. I will leave the implemented features in case
we want to experiment some more. For now the +tofmodels use +zvecs.best
which is just a call to +zvecs.equal_dr which is just a call to linspace.