New and default way ot define the core.
tof.core_mass() and tof.core_radius() refactored to take a single,
required argument, how. how=='byeos' and how=='bypeaks' are kept
as valid options but the new recommended way is to pass a threshold
density. The R_core and M_core properties call the respective method
with how=4500, about the highest density expected with no ice or
rock content.

This is not 100% satisfactory but is much better than the previous
way.

But now we have a different problem: what to make of models that
have a strong density jump but level off below ice/rock densities?