Skip-and-spline solution of figure functions in tof4.
Implemented skip-n-spline speedup in tof4. The idea is to solve the
figure functions (eqs. B12-B15 in Nettelmann 2015) on a subset of
the level surfaces, and spline-interpolate the solution to the rest.
This sounds like cheating but in fact the very smooth nature of the
spheroid shape as a functions of mean radius gives us an advantage.
While the discretized density still needs a fine "grid" in order
for the mass integrals to converge well, the shape functions can
be solved on a much coarser grid.

The implementation includes an additional parameter, sskip, in
tof4.m, with the associated call from TOFPlanet using the value
in obj.opts.splineskip, which is set using tofset(). For now
the default value of sskip=0 retains the original implementation,
with no skipping or splining at all. A value of sskip=1 is used
to verify the algorithm. But in fact a value as high as sskip=32
can be safely used with 2048 density levels to gain a 32-fold
speedup with an error below the tof4 truncation error!

Tested against original calculation with linear density profile and
against index-1-polytrope solution from Wisdom and Hubbard.