Quick overhaul of piecewise-quadratic models.
Instead of forever tweaking the same +tofmodels.tqrep.m
we now have a series of +tofmodels.tripquad_v{x}. For now
we have versions x=1,2,3,4. Version 1 is the old triple
quadratic, using polynomial coefficients. The rest are
variations on the curvature-endpoints parametrization.

Also took the opportunity to move the optional and rarely
used zstrat argument to the end and make it undocumented.