Bug fix:
In relax_to_barotrope() calling renormalize on every iteration to
keep densities reasonable. If we wait till the end of the look the
density values can get out of hand, occasionally becoming negative
and causing a crash. This happens when incompatible barotropes are
assigned, which happens all too often with synthetic barotropes.