MoI calculation improvement.
The old NMoI calculation was a 0-order approximation that only
really worked for spherical planets. Replaced it with a proper
integration of shape. Using Gauss-quad as in CMSPlanet for speed.

Also, replaced dependent prop NMoI and getter with a regular public
method NMoI(). No real good reason, just because.

Also added a sanity check verify_mean_radii() method. We don't
really need it though, sanity is good. This was in the process of
hunting down the cause of differences in MoI calculation in
CMSPlanet and TOFPlanet that turned out to be unrelated. The tof
and cms MoI now agree to ~1e-4 relative, same as J2.