Change default integration of mass and MoI.
Renewed interest in the precision of density integrals led me to
change my mind about the defaults. The mass integral in TOFPlanet
already has several integration options implemented, controlled by
the undocumented opts.masmeth. I changed the default from 'layerz'
to 'trapz'. Since the mass integral is one dimensional (using
spherical shells with radii si) we can easily use the trapezoid
method, and it brings us much better precision.

The MoI integral on the other hand is two-dimensional, unless we
make an explicit approximation of constant density shells and then
integrate only over theta, with gauss-legendre. It appears that we
get faster convergence of MoI value with number of layers when we
use midlayer-densities in this integral. This shows superficial
similarity with the trapezoid rule, but this is an illusion. Still,
the fast convergence makes it tempting to believe that we are
getting a more precise value this way. So TOFPlanet.NMoI now uses
midlayer densities by default, controlled by another undocumented
opts.moimeth (default 'midlayerz'). The midlayer density is also
implemented in tof4.py.