Forsaking fsolve.
Following an email with Nadine I realized that fsolving the An=0
equations might be unnecessary. Instead, iteratively updating each
s_n by setting the corresponding A_n to zero and "solving"
s_n=A_n/S_0 could end up eventually converging to a self-consistent
solution. Turns out, it does, and with only a few more iterations
than we already must use to converge the S_n integrals. Of course
forsaking fsolve gives a huge performance gain.

In fact when fsolve goes away the spline in skip_n_spline_B1215()
becomes the dominant component, for N in the thousands. I find that
for N=4096, say, and if we explicitly bypass the spline call when
nx==N, we barely gain anything from skip-n-spline. For this reason
I now use xlevels=-1 as default.