Squashed commit of branch qrot:
The big idea is to commit to consistently defining a planet
by its unambiguous reference physical quantities. This includes
mass, equatorial radius, and now, rotation period. Thus mrot is
no longer a prime variable and is relegated to a dependant
property with an access method; likewise qrot and wrot. There
will be additional fallout from this switch, to be sure, for now
the main downstream effect and the proximate reason for the change
is a different treatment of barotrope relaxation. obj.relax_to_baro
no longer calls obj.relax_to_HE; it calls tof<n>() directly, going
for FULL convergence each time (slow but turns out to be necessary,
ss_guesses help) then followed by renorm which automatically
recalculates mrot. This is the only way to be completely
consistent with the planet's reference conditions. Indeed with this
procedure, we finally see good agreement with the CLC numbers of
WH16 (if we also use their big G).

The n=1 poly example was of course tweaked, and the constant
density example had to be reworked to use tof<n> since  TOFPlanet
doesn't do mrot-based definitions any more.

Started separate n=1 of N test suite because for better or worse
this is the de facto required benchmark. More work is needed but
it desevres its own branch.

Expect additonal tweaks to come through as fallout from removing
mrot propagates.