Testing two ideas for improving n1ofN convergence, both fail.
First was to improve the order of integration of the hydrostatic
equilibrium equation to obtain possibly more accurate pressure.
Implemented a Simpson integration instead of trapz. Made no
difference. I kept that option in, with an indocumented
opts.prsmeth to switch it on.

Second was to use mid-point pressure in update_density() instead of
pressure on the level surfaces, to mimic what is done in CMS. This
makes things much worse, as expected, and again confirms that mixing
ideas from CMS and TOF is often misleading. Removed it since it's
wrong.