Squashed commit of ss_guesses acceleration.
Keeping and re-using ss on successive calls to relax_to_HE()
while limiting HE iterations results in significant
acceleration of relax_to_rotation(). (Same would be true of
relax_to_barotrope() if/when needed.)

I find that 3-4 HE iterations at a time are a good balance,
based on not-that-many-tests.

The ss and SS going back and forth to/from tof<n> calls (only
ss are used so far) are just kept in self.ss and self.SS as is.
If they are to be used for something other than acceleration
then we should consider flipud-ing them to be consistent with
the density level numbering.

With the addition of the optional ss_guesses there are enough
optional parameters to tof<n> to warrant a switch to **kwargs
signature. This *should* be completely backwards compatible.