Careful with rotation variables!
The only rotation set during set_observables is period.

wrot and mrot are calculated just when needed. They are still
stored in self.wrot and self.mrot because that is so convenient,
but they should be considered private (and will soon be renamed).

Remove mrot and qrot from exporters to avoid ambiguity.