Squashed commit of branch barowork.
TOFPlanet.py now supports barotrope mode. Some changes to TOFPlanet.m
in the process of syncs, including some theoretically not strictly
backwards compatible, but I don't anticipate any issues:

* The J tolerance read from opts.dJTol is now interpreted differently.
  In tp.relax_to* methods it is checked simply against abs(J2 - J2_old),
  under the assumption that J2 is always the most important and most
  precisely known value, and that J2 is always OOM 1e-2 so absolute diff
  is easiest to work with.
* The return value of relax_to_barotrope(), which is never used anyway,
  is changed from elapsed time to iteration count
* The break conditions in relax_to_barotrope() checks also for rotation
  convergence. Just for completeness, as this is almost certainly
  satisfied before dJ of even drho.
* Calculated properties mrot and qrot use fixed ref GM instead of G*M;
  should make no difference since we always renormalize anyway.

The rotating and nonrotating polytrope example polytrope scripts are
slightly tweaked and .py versions added. The output tables show
absolute instead of relative error. As a reminder: it takes N>2^18 for
tof7 to get a J2 truly close to WH16 CLC values, and not even then to
the promised 1e-10 error. For that level of precision it might be
necessary to take a closer look at optimizing for floating-point
arithmetic, but that is a story for another branch.