Refactor TOFPlanet.py::self.baro() to use single pressure argument.
The second radius argument used to support different barotropes in
different layers is potentially inconsistent when radius levels
change slightly during convergence. Better to follow the TP.m
implementation and call different barotropes based on level surface
index, immutable throughout the relaxation. It means providing the
indices that split the barotropes during the definition of self.baro
instead of during the call. For example, a triple polytrope will
define the self.baro with the r_trans and r_core indices
predetermined and fixed.